+++ CHANGELOG.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.2.5] - 2024/10/18
+
+### Added
+
+- Langchain integration
+- local RAG example
+
+### Updated
+
+- Various refactoring and improvements
+
 ## [0.2.4] - 2024/06/04
 
 ### Fixed